+++ .gitignore
@@ -32,3 +32,4 @@
 yarn-debug.log*
 .yarn-integrity
 /coverage/
+.history/
