+++ app/javascript/packs/application.js
@@ -1,4 +1,4 @@
 import WebpackerReact from "webpacker-react";
-import TaskBoard from "./components/TaskBoard";
+import TaskBoard from "components/TaskBoard";
 
 WebpackerReact.setup({ TaskBoard });
