+++ app/javascript/services/FetchService.js
